clients/checkout: split and refactor components
